Add CSS resources processing
