+++ jupyter-lib/shared-compiler/src/main/kotlin/org/jetbrains/kotlinx/jupyter/libraries/JsLibraryResourcesProcessor.kt
@@ -1,0 +1,123 @@
+package org.jetbrains.kotlinx.jupyter.libraries
+
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
+import org.intellij.lang.annotations.Language
+import org.jetbrains.kotlinx.jupyter.api.Code
+import org.jetbrains.kotlinx.jupyter.api.libraries.LibraryResource
+import org.jetbrains.kotlinx.jupyter.api.libraries.ResourcePathType
+import org.jetbrains.kotlinx.jupyter.config.getLogger
+import java.io.File
+
+class JsLibraryResourcesProcessor : LibraryResourcesProcessor {
+    private var outputCounter = 0
+
+    private fun loadResourceAsText(resource: LibraryResource, classLoader: ClassLoader): List<ScriptModifierFunctionGenerator> {
+        return resource.locations.map { location ->
+            val pathString = location.path
+            when (location.type) {
+                ResourcePathType.URL -> {
+                    URLScriptModifierFunctionGenerator(pathString)
+                }
+                ResourcePathType.LOCAL_PATH -> {
+                    val file = File(pathString)
+                    logger.debug("Resolving resource file: ${file.absolutePath}")
+                    CodeScriptModifierFunctionGenerator(file.readText())
+                }
+                ResourcePathType.CLASSPATH_PATH -> {
+                    CodeScriptModifierFunctionGenerator(classLoader.getResource(pathString)?.readText().orEmpty())
+                }
+            }
+        }
+    }
+
+    override fun wrapLibrary(resource: LibraryResource, classLoader: ClassLoader): String {
+        val resourceName = resource.name
+        val elementId = "kotlin_out_$outputCounter"
+        ++outputCounter
+
+        val generators = loadResourceAsText(resource, classLoader)
+        val jsScriptModifiers = generators.joinToString(",\n", "[", "]") {
+            it.getScriptText()
+        }
+
+        @Language("js")
+        val wrapper = """
+            if(!window.kotlinQueues) {
+                window.kotlinQueues = {};
+            }
+            if(!window.kotlinQueues.$resourceName) {
+                var resQueue = [];
+                window.kotlinQueues.$resourceName = resQueue;
+                window.call_$resourceName = function(f) {
+                    resQueue.push(f);
+                }
+            }
+            (function (){
+                var modifiers = $jsScriptModifiers
+                var e = document.getElementById("$elementId");
+                modifiers.forEach(function (gen) {
+                    var script = document.createElement("script");
+                    gen(script)
+                    script.type = "text/javascript";
+                    script.onload = function() {
+                        window.call_$resourceName = function(f) {f();};
+                        window.kotlinQueues.$resourceName.forEach(function(f) {f();});
+                        window.kotlinQueues.$resourceName = [];
+                    };
+                    script.onerror = function() {
+                        window.call_$resourceName = function(f) {};
+                        window.kotlinQueues.$resourceName = [];
+                        var div = document.createElement("div");
+                        div.style.color = 'darkred';
+                        div.textContent = 'Error loading resource $resourceName';
+                        document.getElementById("$elementId").appendChild(div);
+                    };
+                    
+                    e.appendChild(script);
+                })
+            })()
+        """.trimIndent()
+
+        // language=html
+        return """
+            <div id="$elementId"/>
+            <script type="text/javascript">
+                $wrapper
+            </script>
+        """.trimIndent()
+    }
+
+    private interface ScriptModifierFunctionGenerator {
+        fun getScriptText(): String
+    }
+
+    private class CodeScriptModifierFunctionGenerator(
+        val code: Code
+    ) : ScriptModifierFunctionGenerator {
+        override fun getScriptText(): String {
+            val escapedCode = Json.encodeToString(code)
+            // language=js
+            return """
+                (function(script) {
+                    script.textContent = $escapedCode
+                })
+            """.trimIndent()
+        }
+    }
+
+    private class URLScriptModifierFunctionGenerator(
+        private val url: String
+    ) : ScriptModifierFunctionGenerator {
+        override fun getScriptText(): String {
+            // language=js
+            return """
+                (function(script) {
+                    script.src = "$url"
+                })
+            """.trimIndent()
+        }
+    }
+
+    private val logger = getLogger(JsLibraryResourcesProcessor::class.simpleName!!)
+}
